Format segnum

+++ editor/centers.cpp
@@ -200,7 +200,7 @@ int centers_dialog_handler(UI_DIALOG *dlg, d_event *event, centers_dialog *c)
 //		int	i;
 //		char	temp_text[CENTER_STRING_LENGTH];
 	
-		ui_dprintf_at( dlg, 12, 6, "Seg: %3hu", Cursegp-Segments );
+		ui_dprintf_at( dlg, 12, 6, "Seg: %3hu", static_cast<unsigned short>(Cursegp-Segments) );
 
 //		for (i=0; i<CENTER_STRING_LENGTH; i++)
 //			temp_text[i] = ' ';
